refactor: put local components into same named folder
