+++ src/config/api.ts
@@ -1,0 +1,5 @@
+const API_ENDPOINTS = {
+  USER: './api/user',
+};
+
+export default API_ENDPOINTS;
