+++ src/pages/MainPage/components/Pagination/Pagination.module.scss
