+++ src/pages/MainPage/components/Pagination/index.ts
