+++ src/pages/RootPage/components/Header/Logo/Logo.module.scss
