+++ src/pages/RootPage/components/Header/Logo/index.ts
