+++ src/pages/RootPage/components/Header/NavBar/NavBar.module.scss
